+++ internal/db/db_test.go
@@ -5,7 +5,7 @@ import (
 	"log/slog"
 	"testing"
 
-	"github.com/nathanhollows/Rapua/v7/db"
+	"github.com/nathanhollows/Rapua/v7/internal/db"
 	"github.com/stretchr/testify/require"
 )