+++ internal/services/monthly_credit_topup.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/nathanhollows/Rapua/v7/config"
-	"github.com/nathanhollows/Rapua/v7/db"
+	"github.com/nathanhollows/Rapua/v7/internal/db"
 	"github.com/nathanhollows/Rapua/v7/models"
 	"github.com/uptrace/bun"
 )