+++ internal/services/teams_service.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/nathanhollows/Rapua/v7/db"
 	"github.com/nathanhollows/Rapua/v7/helpers"
+	"github.com/nathanhollows/Rapua/v7/internal/db"
 	"github.com/nathanhollows/Rapua/v7/models"
 	"github.com/nathanhollows/Rapua/v7/repositories"
 	"github.com/uptrace/bun"