+++ internal/services/utils_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/brianvoe/gofakeit/v7"
-	"github.com/nathanhollows/Rapua/v7/db"
+	"github.com/nathanhollows/Rapua/v7/internal/db"
 	"github.com/nathanhollows/Rapua/v7/internal/migrations"
 	"github.com/nathanhollows/Rapua/v7/models"
 	"github.com/uptrace/bun"